refactor(server): restructure mongodb dao implementation to match mysql patterns by splitting grouped daos, centralizing mongo helpers, normalizing exports, and updating factory/routes/tests
