+++ packages/server/src/dao/mongo/query.mongo.dao.ts
@@ -1,0 +1,1 @@
+export { executeMongoQuery, executeMongoQuery as executeQuery } from "./query.dao.js";
