+++ packages/server/src/dao/mongo/schema.mongo.dao.ts
@@ -1,0 +1,1 @@
+export { getMongoDatabaseSchema } from "./schema.dao.js";
