+++ packages/server/tests/routes/mongodb/query.routes.mongodb.test.ts
@@ -2,10 +2,11 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { HTTPException } from "hono/http-exception";
 
 import { createServer } from "@/utils/create-server.js";
-import * as mongoQueryDao from "@/dao/mongo/query.dao.js";
+import * as mongoQueryDao from "@/dao/mongo/query.mongo.dao.js";
 
 // Mock MongoDB query DAO
-vi.mock("@/dao/mongo/query.dao.js", () => ({
+vi.mock("@/dao/mongo/query.mongo.dao.js", () => ({
+	executeQuery: vi.fn(),
 	executeMongoQuery: vi.fn(),
 }));
 
@@ -53,7 +54,7 @@ describe("Query Routes (MongoDB)", () => {
 				duration: 8.4,
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "users",
@@ -71,7 +72,7 @@ describe("Query Routes (MongoDB)", () => {
 			expect(res.status).toBe(200);
 			const json = await res.json();
 			expect(json.data).toEqual(mockResult);
-			expect(mongoQueryDao.executeMongoQuery).toHaveBeenCalledWith({
+			expect(mongoQueryDao.executeQuery).toHaveBeenCalledWith({
 				query,
 				db: "testdb",
 			});
@@ -88,7 +89,7 @@ describe("Query Routes (MongoDB)", () => {
 				duration: 15.2,
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "orders",
@@ -119,7 +120,7 @@ describe("Query Routes (MongoDB)", () => {
 				message: "OK",
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "users",
@@ -148,7 +149,7 @@ describe("Query Routes (MongoDB)", () => {
 				message: "OK",
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "products",
@@ -180,7 +181,7 @@ describe("Query Routes (MongoDB)", () => {
 				message: "OK (1 modified)",
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "users",
@@ -209,7 +210,7 @@ describe("Query Routes (MongoDB)", () => {
 				message: "OK (5 modified)",
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "users",
@@ -238,7 +239,7 @@ describe("Query Routes (MongoDB)", () => {
 				message: "OK",
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "users",
@@ -264,7 +265,7 @@ describe("Query Routes (MongoDB)", () => {
 				message: "OK",
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "sessions",
@@ -292,7 +293,7 @@ describe("Query Routes (MongoDB)", () => {
 				message: "OK",
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const query = JSON.stringify({
 				collection: "users",
@@ -312,7 +313,7 @@ describe("Query Routes (MongoDB)", () => {
 		});
 
 		it("returns 400 for empty query string", async () => {
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockRejectedValue(
+			vi.mocked(mongoQueryDao.executeQuery).mockRejectedValue(
 				new HTTPException(400, { message: "Query is required" }),
 			);
 
@@ -326,7 +327,7 @@ describe("Query Routes (MongoDB)", () => {
 		});
 
 		it("returns 400 for invalid JSON query payload", async () => {
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockRejectedValue(
+			vi.mocked(mongoQueryDao.executeQuery).mockRejectedValue(
 				new HTTPException(400, { message: "Mongo query must be valid JSON" }),
 			);
 
@@ -340,7 +341,7 @@ describe("Query Routes (MongoDB)", () => {
 		});
 
 		it("returns 400 when collection or operation is missing", async () => {
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockRejectedValue(
+			vi.mocked(mongoQueryDao.executeQuery).mockRejectedValue(
 				new HTTPException(400, {
 					message: "Mongo query must include collection and operation",
 				}),
@@ -356,7 +357,7 @@ describe("Query Routes (MongoDB)", () => {
 		});
 
 		it("returns 400 for unsupported operation", async () => {
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockRejectedValue(
+			vi.mocked(mongoQueryDao.executeQuery).mockRejectedValue(
 				new HTTPException(400, { message: "Unsupported Mongo operation" }),
 			);
 
@@ -394,7 +395,7 @@ describe("Query Routes (MongoDB)", () => {
 		});
 
 		it("returns 503 on connection failure", async () => {
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockRejectedValue(
+			vi.mocked(mongoQueryDao.executeQuery).mockRejectedValue(
 				new Error("connect ECONNREFUSED 127.0.0.1:27017"),
 			);
 
@@ -410,7 +411,7 @@ describe("Query Routes (MongoDB)", () => {
 		});
 
 		it("returns 503 on connection timeout", async () => {
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockRejectedValue(
+			vi.mocked(mongoQueryDao.executeQuery).mockRejectedValue(
 				new Error("timeout expired"),
 			);
 
@@ -433,7 +434,7 @@ describe("Query Routes (MongoDB)", () => {
 				duration: 12.7,
 			};
 
-			vi.mocked(mongoQueryDao.executeMongoQuery).mockResolvedValue(mockResult);
+			vi.mocked(mongoQueryDao.executeQuery).mockResolvedValue(mockResult);
 
 			const res = await app.request("/mongodb/query?db=testdb", {
 				method: "POST",
